+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ScottPlot 4.1.61
 _not yet published on NuGet..._
 * Axis: Throw exception immediately upon setting invalid axis limits (#2327) _Thanks @mjpz_
+* Heatmap: Added support for transparent single-color heatmaps (#2336) _Thanks @bukkideme_
 
 ## ScottPlot 4.1.60
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2022-12-23_